protected routes with jwt_required() and set up relationship between users and notes

+++ main.py
@@ -1,7 +1,7 @@
 from flask import Flask, request, jsonify
 import bcrypt
 from models import db, Note, User
-from flask_jwt_extended import JWTManager, create_access_token, jwt_required
+from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt_identity
 
 app = Flask(__name__)
 app.config["SQLALCHEMY_DATABASE_URI"] = 'sqlite:///notes.db'
@@ -16,23 +16,31 @@ with app.app_context():
 @app.route('/api/notes', methods=['GET'])
 @jwt_required()
 def get_notes():
-    notes = db.session.execute(db.select(Note)).scalars()
+    user_id = get_jwt_identity()
+    notes = db.session.execute(db.select(Note).where(Note.user_id == int(user_id))).scalars()
     notes_data = [note.to_dict() for note in notes]
     return jsonify(notes_data)
 
 @app.route('/api/notes', methods=['POST'])
+@jwt_required()
 def add_note():
     data = request.get_json()
-    new_note = Note(title=data.get("title"), content=data.get("content"))
+    user_id = get_jwt_identity()
+    new_note = Note(title=data.get("title"), content=data.get("content"), user_id=int(user_id))
     db.session.add(new_note)
     db.session.commit()
     return jsonify(new_note.to_dict()), 201
 
 @app.route('/api/notes/<int:note_id>', methods=['PUT'])
+@jwt_required()
 def update_note(note_id):
     note = db.session.get(Note, note_id)
-    
+    user_id = int(get_jwt_identity())
+
     if note:
+        if note.user_id != user_id:
+            return jsonify("error: bad credentials"), 401
+        
         updated_content = request.get_json()
         note.title = updated_content.get("title", note.title)
         note.content = updated_content.get("content", note.content)
@@ -43,10 +51,15 @@ def update_note(note_id):
     return jsonify("Error: resource not found"), 404
 
 @app.route('/api/notes/<int:note_id>', methods=['DELETE'])
+@jwt_required()
 def delete_note(note_id):
     note = db.session.get(Note, note_id)
+    user_id = int(get_jwt_identity())
 
     if note:
+        if note.user_id != user_id:
+            return jsonify("error: bad credentials"), 401
+        
         db.session.delete(note)
         db.session.commit()
         return "", 204
